Convert access methods to repos
